Refactored to get Scala 2.11 to compile (non-tests)
