+++ kernel/src/main/scala/org/apache/toree/kernel/protocol/v5/handler/CommMsgHandler.scala
@@ -26,7 +26,6 @@ import play.api.data.validation.ValidationError
 import play.api.libs.json.JsPath
 
 import scala.concurrent.Future
-import scala.concurrent.future
 import scala.concurrent.ExecutionContext.Implicits.global
 
 /**
@@ -41,7 +40,7 @@ class CommMsgHandler(
   commStorage: CommStorage
 ) extends BaseHandler(actorLoader) with MessageLogSupport
 {
-  override def process(kernelMessage: KernelMessage): Future[_] = future {
+  override def process(kernelMessage: KernelMessage): Future[_] = Future {
     logKernelMessageAction("Initiating Comm Msg for", kernelMessage)
 
     val kmBuilder = KMBuilder().withParent(kernelMessage)
