+++ kernel/src/main/scala/org/apache/toree/kernel/protocol/v5/handler/KernelInfoRequestHandler.scala
@@ -34,7 +34,7 @@ class KernelInfoRequestHandler(actorLoader: ActorLoader)
 {
   def process(kernelMessage: KernelMessage): Future[_] = {
     import scala.concurrent.ExecutionContext.Implicits.global
-    future {
+    Future {
       logger.debug("Sending kernel info reply message")
 
       val kernelInfo = SparkKernelInfo
