--- scala-interpreter/src/main/scala/org/apache/toree/kernel/interpreter/scala/SparkIMainProducerLike.scala
@@ -1,46 +1,0 @@
-/*
- *  Licensed to the Apache Software Foundation (ASF) under one or more
- *  contributor license agreements.  See the NOTICE file distributed with
- *  this work for additional information regarding copyright ownership.
- *  The ASF licenses this file to You under the Apache License, Version 2.0
- *  (the "License"); you may not use this file except in compliance with
- *  the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License
- */
-
-package org.apache.toree.kernel.interpreter.scala
-
-import org.apache.spark.repl.SparkIMain
-
-import scala.tools.nsc.interpreter.JPrintWriter
-import scala.tools.nsc.{Settings, interpreter}
-
-trait SparkIMainProducerLike {
-  /**
-   * Constructs a new instance of SparkIMain.
-   *
-   * @param settings The settings associated with the SparkIMain instance
-   * @param out The output writer associated with the SparkIMain instance
-   *
-   * @return The new SparkIMain instance
-   */
-  def newSparkIMain(settings: Settings, out: interpreter.JPrintWriter): SparkIMain
-}
-
-trait StandardSparkIMainProducer extends SparkIMainProducerLike {
-  override def newSparkIMain(
-    settings: Settings, out: JPrintWriter
-  ): SparkIMain = {
-    val s = new SparkIMain(settings, out)
-    s.initializeSynchronous()
-
-    s
-  }
-}
